Added two instances to IDO class
It is the first time I add instances. I hope I did it correctly.
The IDs of the IDOs have to be changed. For now they are 1 and 2 respectively. If I remember, Elizabeth or someone else talked about more specific IDs for IDOs in some message. I am going to look for information now.