Added 1 instance to CCI class and changed name of CCI class to Cross-Cutting Objective
I added the instance Adaptation to and Mitigation of Climate Change Through Agricultural Research.

I thought whether I should add Climate Change as hasExactSynonim but I decided not to do that. CGIAR uses Climate Change as the label for this objective. However, the label Climate Change already exists in the ontology (I believe) and it refers to another entitiy.

I change the name of the CCI class to "cross-cutting objective" because what we list are objective and not issues, properly speaking. I added "cross-cutting issue" as hasExactSynonim.